Even much cleaner C interface
C interface completely independent from C++ interface.
